singedとunsginedの変数/定数の比較を行わないようにした
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/trunk@7724 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ teraterm/teraterm/vtwin.cpp
@@ -6380,7 +6380,7 @@ LRESULT CVTWindow::OnDpiChanged(WPARAM wp, LPARAM)
 
 	// Šm”F
 	const RECT *NewRect = &NewWindowRect[0];
-	for (int i=0; i < _countof(NewWindowRect); i++) {
+	for (size_t i=0; i < _countof(NewWindowRect); i++) {
 		const RECT *r = &NewWindowRect[i];
 		HMONITOR hMonitor = pMonitorFromRect(r, MONITOR_DEFAULTTONULL);
 		UINT dpiX;
